refactor: Moved DisplayNotid to own file as component
